move testapp to examples folder
